refactor(types): group platform/host types under common/types/platform
