+++ packages/desktop/src/common/types/platform/electron.ts
