+++ packages/desktop/src/common/types/platform/fileSnapshot.ts
