+++ packages/desktop/src/renderer/components/agent/AgentModeSelector.tsx
@@ -6,7 +6,7 @@
 
 import { ipcBridge } from '@/common';
 import { configService } from '@/common/config/configService';
-import type { AcpSessionConfigOption } from '@/common/types/acpTypes';
+import type { AcpSessionConfigOption } from '@/common/types/platform/acpTypes';
 import { getAgentModes, supportsModeSwitch, type AgentModeOption } from '@/renderer/utils/model/agentModes';
 import { useLayoutContext } from '@/renderer/hooks/context/LayoutContext';
 import { AgentLogoIcon } from './AgentBadge';
